Move env tests into trait
